fix(go): avoid sqlite race test collision
