+++ internal/autosearch/engine.go
@@ -17,7 +17,7 @@ import (
 	"github.com/ebenderooock/loom/internal/auditlog"
 	"github.com/ebenderooock/loom/internal/customformats"
 	"github.com/ebenderooock/loom/internal/downloads"
-	torrentdl "github.com/ebenderooock/loom/internal/downloads/torrent"
+	"github.com/ebenderooock/loom/internal/downloads/torrentutil"
 	"github.com/ebenderooock/loom/internal/indexers"
 	"github.com/ebenderooock/loom/internal/movies"
 	"github.com/ebenderooock/loom/internal/parser"
@@ -1426,7 +1426,7 @@ func buildDownloadRequest(res *indexers.Result) downloads.AddRequest {
 	if res.MagnetURI != "" {
 		req.Magnet = res.MagnetURI
 	} else if res.Infohash != "" {
-		req.Magnet = torrentdl.BuildPublicMagnet(res.Infohash, res.Title)
+		req.Magnet = torrentutil.BuildPublicMagnet(res.Infohash, res.Title)
 	}
 	if res.Link != "" {
 		// Some indexers (e.g. EZTV) store the magnet URI in the Link field
